requirements.txt
# =============================================================================
# RAG Evaluation Pipeline — Dependencies
# =============================================================================
# Install with:  pip install -r requirements.txt
#
# Python 3.11+ is required.
#
# API keys (set in .env or environment):
#   OPENAI_API_KEY    — required if using openai provider
#   ANTHROPIC_API_KEY — required if using anthropic provider
# =============================================================================

# ---------------------------------------------------------------------------
# LangChain core + integrations
# ---------------------------------------------------------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4

# ---------------------------------------------------------------------------
# Embedding models (local / HuggingFace)
# ---------------------------------------------------------------------------
sentence-transformers>=3.0.0
torch>=2.0.0               # required by sentence-transformers
transformers>=4.40.0       # required by sentence-transformers

# ---------------------------------------------------------------------------
# Vector databases
# ---------------------------------------------------------------------------
faiss-cpu>=1.8.0           # Change to faiss-gpu if GPU available
chromadb>=0.5.0

# ---------------------------------------------------------------------------
# LLM providers
# ---------------------------------------------------------------------------
openai>=1.50.0
anthropic>=0.40.0

# ---------------------------------------------------------------------------
# Document loading
# ---------------------------------------------------------------------------
pypdf>=4.0.0               # PDF support

# ---------------------------------------------------------------------------
# Evaluation framework
# ---------------------------------------------------------------------------
ragas>=0.2.0

# ---------------------------------------------------------------------------
# Data processing and analysis
# ---------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.26.0

# ---------------------------------------------------------------------------
# Visualisation
# ---------------------------------------------------------------------------
matplotlib>=3.8.0
seaborn>=0.13.0

# ---------------------------------------------------------------------------
# Utilities
# ---------------------------------------------------------------------------
python-dotenv>=1.0.0       # .env file support
tqdm>=4.66.0               # progress bars
pydantic>=2.0.0            # data validation (used by LangChain/RAGAS)
tiktoken>=0.7.0            # token counting for OpenAI models

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
